update the testcases

+++ user/src/oscomp/unlink.c
@@ -19,22 +19,31 @@ int test_unlink()
         int fd, ret;
 
         fd = open(fname, O_CREATE | O_WRONLY);
-        assert(fd > 0);
+        if (fd < 0) {
+                printf("ERROR: open syscall not implemented or failed (returned %d)\n", fd);
+                TEST_FAIL(__func__);
+                return;
+        }
         close(fd);
 
         // unlink test
         ret = unlink(fname);
-        assert(ret == 0);
+        if (ret < 0) {
+                printf("ERROR: unlink syscall not implemented or failed (returned %d)\n", ret);
+                TEST_FAIL(__func__);
+                return;
+        }
+
         fd = open(fname, O_RDONLY);
         if (fd < 0) {
                 printf("  unlink success!\n");
+                TEST_PASS(__func__);
         } else {
                 printf("  unlink error!\n");
                 close(fd);
+                TEST_FAIL(__func__);
         }
         // It's Ok if you don't delete the inode and data blocks.
-
-        TEST_END(__func__);
 }
 
 int main(void)
